services/users/func-try-user-disable-automation/requirements.txt
python-dotenv==1.0.0
pytest
black
azure-keyvault-secrets
azure-identity
requests
pytz
azure-functions